+++ services/auth/lib/local.js
@@ -3,7 +3,7 @@ const handlers = require('./handlers');
 const helpers = require('../../../lib/modules/responseHelpers');
 const state = require('./state');
 const bcrypt = require('bcryptjs');
-const { getUsersCollection } = require('./modules/collectionNames');
+const { getUsersCollection } = require('./modules/authCollections');
 const { isEmailValid } = require('./handlers');
 const editURL = require('edit-url');
 const debug = require('debug')('campsi:auth:local');
