+++ services/docs/lib/services/document.js
@@ -7,7 +7,7 @@ const { ObjectId } = require('mongodb');
 
 const createError = require('http-errors');
 const { getDocumentLockServiceOptions } = require('../modules/serviceOptions');
-const { getUsersCollection } = require('../../../auth/lib/modules/collectionNames');
+const { getUsersCollection } = require('../../../auth/lib/modules/authCollections');
 const { paginateQuery } = require('../../../../lib/modules/paginateCursor');
 
 // Helper functions
