+++ test/auth/collection-name-default.js
@@ -4,14 +4,11 @@ process.env.NODE_CONFIG_DIR = './test/config';
 
 // Require the dev-dependencies
 const debug = require('debug')('campsi:test');
-const { MongoClient } = require('mongodb');
 const config = require('config');
 const chai = require('chai');
 const CampsiServer = require('campsi');
-const mongoUriBuilder = require('mongo-uri-builder');
 const chaiHttp = require('chai-http');
 const format = require('string-format');
-const { getUsersCollectionName, getSessionCollectionName } = require('../../services/auth/lib/modules/collectionNames');
 const AuthService = require('../../services/auth/lib');
 const { emptyDatabase } = require('../helpers/emptyDatabase');
 const expect = chai.expect;
